@@ -1,9 +1,9 @@
-const formatStylish = require('./stylish');
-const formatPlain = require('./plain');
-const formatJson = require('./json');
+import formatStylish from './stylish.js';
+import formatPlain from './plain.js';
+import formatJson from './json.js';
 
-module.exports = {
-    formatStylish,
-    formatPlain,
-    formatJson
+export {
+  formatStylish,
+  formatPlain,
+  formatJson,
 };
