@@ -1,18 +1,17 @@
-const fs = require('fs');
-const path = require('path');
-const yaml = require('js-yaml');
+import fs from 'fs';
+import path from 'path';
+import yaml from 'js-yaml';
 
 function parseFile(filepath) {
-    const ext = path.extname(filepath);
-    const content = fs.readFileSync(filepath, 'utf8');
-    
-    if (ext === '.json') {
-        return JSON.parse(content);
-    } else if (ext === '.yml' || ext === '.yaml') {
-        return yaml.load(content);
-    } else {
-        throw new Error(`Unsupported file format: ${ext}`);
-    }
+  const ext = path.extname(filepath);
+  const content = fs.readFileSync(filepath, 'utf8');
+
+  if (ext === '.json') {
+    return JSON.parse(content);
+  } if (ext === '.yml' || ext === '.yaml') {
+    return yaml.load(content);
+  }
+  throw new Error(`Unsupported file format: ${ext}`);
 }
 
-module.exports = parseFile;
+export default parseFile;
